// src/generated/ast.ts
export type CoclContext = CoclDerived | CoclReferenceFilter | CoclValidation;

export const CoclContext = 'CoclContext';
// ...
export type ContextDeclaration = ClassifierContext | CoclContext | OperationContext | PropertyContext;

export const ContextDeclaration = 'ContextDeclaration';

export function isContextDeclaration(item: unknown): item is ContextDeclaration {
    return reflection.isInstance(item, ContextDeclaration);
}

export type Expression = ArrowExpression | BinaryExpression | BooleanLiteral | CollectionLiteral | DotExpression | IfExpression | IntegerLiteral | InvalidLiteral | LetExpression | NullLiteral | RealLiteral | SelfExpression | StringLiteral | TupleLiteral | TypeExpLiteral | UnaryExpression | VariableExpression;

export const Expression = 'Expression';
// ...
export interface ArrowExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'ArrowExpression';
    arguments: Array<Expression | IteratorBody>;
    operation: string;
    source: Expression;
}

export const ArrowExpression = 'ArrowExpression';

export function isArrowExpression(item: unknown): item is ArrowExpression {
    return reflection.isInstance(item, ArrowExpression);
}

export interface BinaryExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'BinaryExpression';
    left: Expression;
    op: '*' | '+' | '-' | '/' | '<' | '<=' | '<>' | '=' | '>' | '>=' | 'and' | 'div' | 'implies' | 'mod' | 'or' | 'xor';
    right: Expression;
}

export const BinaryExpression = 'BinaryExpression';

export function isBinaryExpression(item: unknown): item is BinaryExpression {
    return reflection.isInstance(item, BinaryExpression);
}

export interface BodyCondition extends langium.AstNode {
    readonly $container: OperationContext;
    readonly $type: 'BodyCondition';
    expression: Expression;
    name?: string;
}

export const BodyCondition = 'BodyCondition';

export function isBodyCondition(item: unknown): item is BodyCondition {
    return reflection.isInstance(item, BodyCondition);
}

export interface BooleanLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'BooleanLiteral';
    value: 'false' | 'true';
}

export const BooleanLiteral = 'BooleanLiteral';

export function isBooleanLiteral(item: unknown): item is BooleanLiteral {
    return reflection.isInstance(item, BooleanLiteral);
}

export interface ClassifierContext extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'ClassifierContext';
    constraints: Array<ClassifierConstraint>;
    type: QualifiedName;
}

export const ClassifierContext = 'ClassifierContext';

export function isClassifierContext(item: unknown): item is ClassifierContext {
    return reflection.isInstance(item, ClassifierContext);
}

export interface CoclDerived extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'CoclDerived';
    expression: Expression;
    path: QualifiedName;
}

export const CoclDerived = 'CoclDerived';

export function isCoclDerived(item: unknown): item is CoclDerived {
    return reflection.isInstance(item, CoclDerived);
}

export interface CoclReferenceFilter extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'CoclReferenceFilter';
    expression: Expression;
    path: QualifiedName;
}

export const CoclReferenceFilter = 'CoclReferenceFilter';

export function isCoclReferenceFilter(item: unknown): item is CoclReferenceFilter {
    return reflection.isInstance(item, CoclReferenceFilter);
}

export interface CoclValidation extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'CoclValidation';
    expression: Expression;
    path: QualifiedName;
}

export const CoclValidation = 'CoclValidation';

export function isCoclValidation(item: unknown): item is CoclValidation {
    return reflection.isInstance(item, CoclValidation);
}

export interface CollectionItem extends langium.AstNode {
    readonly $container: CollectionLiteral;
    readonly $type: 'CollectionItem';
    first: Expression;
    last?: Expression;
}

export const CollectionItem = 'CollectionItem';

export function isCollectionItem(item: unknown): item is CollectionItem {
    return reflection.isInstance(item, CollectionItem);
}

export interface CollectionLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'CollectionLiteral';
    items: Array<CollectionItem>;
    kind: CollectionKind;
}

export const CollectionLiteral = 'CollectionLiteral';
// ...
export const DefExpression = 'DefExpression';
// ...
export interface DotExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'DotExpression';
    arguments: Array<Expression>;
    feature: string;
    source: Expression;
}

export const DotExpression = 'DotExpression';

export function isDotExpression(item: unknown): item is DotExpression {
    return reflection.isInstance(item, DotExpression);
}

export interface IfExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'IfExpression';
    condition: Expression;
    elseExpression: Expression;
    thenExpression: Expression;
}

export const IfExpression = 'IfExpression';
// ...
export interface IntegerLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'IntegerLiteral';
    value: number;
}

export const IntegerLiteral = 'IntegerLiteral';

export function isIntegerLiteral(item: unknown): item is IntegerLiteral {
    return reflection.isInstance(item, IntegerLiteral);
}

export interface InvalidLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'InvalidLiteral';
}

export const InvalidLiteral = 'InvalidLiteral';
// ...
export interface LetExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'LetExpression';
    body: Expression;
    name: string;
    type?: TypeExpression;
    value: Expression;
}

export const LetExpression = 'LetExpression';

export function isLetExpression(item: unknown): item is LetExpression {
    return reflection.isInstance(item, LetExpression);
}

export interface NullLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'NullLiteral';
}

export const NullLiteral = 'NullLiteral';

export function isNullLiteral(item: unknown): item is NullLiteral {
    return reflection.isInstance(item, NullLiteral);
}

export interface OclDocument extends langium.AstNode {
    readonly $type: 'OclDocument';
    contexts: Array<ContextDeclaration>;
}

export const OclDocument = 'OclDocument';

export function isOclDocument(item: unknown): item is OclDocument {
    return reflection.isInstance(item, OclDocument);
}

export interface OperationContext extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'OperationContext';
    conditions: Array<OperationCondition>;
    parameters: Array<Parameter>;
    path: QualifiedName;
    returnType?: TypeExpression;
}

export const OperationContext = 'OperationContext';
// ...
export interface PropertyContext extends langium.AstNode {
    readonly $container: OclDocument;
    readonly $type: 'PropertyContext';
    constraints: Array<PropertyConstraint>;
    path: QualifiedName;
    propertyType: TypeExpression;
}

export const PropertyContext = 'PropertyContext';

export function isPropertyContext(item: unknown): item is PropertyContext {
    return reflection.isInstance(item, PropertyContext);
}

export interface RealLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'RealLiteral';
    value: number;
}

export const RealLiteral = 'RealLiteral';

export function isRealLiteral(item: unknown): item is RealLiteral {
    return reflection.isInstance(item, RealLiteral);
}

export interface SelfExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'SelfExpression';
}

export const SelfExpression = 'SelfExpression';
// ...
export interface StringLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'StringLiteral';
    value: string;
}

export const StringLiteral = 'StringLiteral';

export function isStringLiteral(item: unknown): item is StringLiteral {
    return reflection.isInstance(item, StringLiteral);
}

export interface TupleLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'TupleLiteral';
    parts: Array<TuplePart>;
}

export const TupleLiteral = 'TupleLiteral';
// ...
export interface TypeExpLiteral extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'TypeExpLiteral';
    feature: string;
    type: QualifiedName;
}

export const TypeExpLiteral = 'TypeExpLiteral';

export function isTypeExpLiteral(item: unknown): item is TypeExpLiteral {
    return reflection.isInstance(item, TypeExpLiteral);
}

export interface UnaryExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'UnaryExpression';
    op: '-' | 'not';
    operand: Expression;
}

export const UnaryExpression = 'UnaryExpression';

export function isUnaryExpression(item: unknown): item is UnaryExpression {
    return reflection.isInstance(item, UnaryExpression);
}

export interface VariableExpression extends langium.AstNode {
    readonly $container: ArrowExpression | BinaryExpression | BodyCondition | CoclDerived | CoclReferenceFilter | CoclValidation | CollectionItem | DefExpression | DeriveConstraint | DotExpression | IfExpression | InitConstraint | InvariantConstraint | IteratorBody | LetExpression | PostCondition | PreCondition | TuplePart | UnaryExpression;
    readonly $type: 'VariableExpression';
    name: QualifiedName;
}

export const VariableExpression = 'VariableExpression';